ml_project/requirements.txt
numpy==1.20.2
pandas==1.2.4
scikit-learn==0.24.1
scipy==1.6.2
pytest==6.2.3
pytest-cov==2.11.1
coverage==5.5
hypothesis==6.10.0
PyYAML==5.4.1
marshmallow==3.11.1
marshmallow-dataclass==8.4.1
six==1.15.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
click==7.1.2
pandas-profiling==2.13.0
attrs==20.3.0
certifi==2020.12.5
cloudpickle==1.6.0
importlib-metadata==4.0.1
iniconfig==1.1.1
joblib==1.0.1
mypy-extensions==0.4.3
packaging==20.9
pluggy==0.13.1
py==1.10.0
pydantic==1.8.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-stdnum==1.16
pytz==2021.1
sortedcontainers==2.3.0
threadpoolctl==2.1.0
toml==0.10.2
zipp==3.4.1
